# [musicdb] improves performance for querying the album view