This will improves the overall performance for queries which select from `albumview`. 
I've dropped the group by from the album view because it dramatically slow down each related query.
Further more I've optimized the albums by artist where condition to use a sub-select instead of group by + join + or, which also reduce the query time.

Related forum thread: 
http://forum.kodi.tv/showthread.php?tid=223731

Some execution times before and after.
```
SELECT albumview.* FROM albumview JOIN song ON song.idAlbum = albumview.idAlbum JOIN song_artist ON song.idSong = song_artist.idSong JOIN album_artist ON albumview.idAlbum = album_artist.idAlbum WHERE song_artist.idArtist = 779 OR  album_artist.idArtist = 779 GROUP BY albumview.idAlbum
```
SQLite: 1306ms
MySQL/MariaDB: 4466ms

```
SELECT albumview.* FROM albumview  WHERE (EXISTS (SELECT 1 FROM song JOIN song_artist ON song.idSong = song_artist.idSong WHERE song.idAlbum = albumview.idAlbum AND song_artist.idArtist = 780 ) OR EXISTS (SELECT 1 FROM album_artist WHERE album_artist.idAlbum = albumview.idAlbum AND album_artist.idArtist = 780))
```
SQLite: 1ms
MySQL/MariaDB: 91ms

@mkortstiege mind taking a look.

@sopparus thanks for sharing your mysql/mariadb server ;)