@Montellese 2015-04-13
Please use `(NOT) EXISTS` instead of `(NOT) IN` as it is further optimized to stop as soon as the condition is met instead of executing the full subquery.