@xhaggi 2015-04-13
In that case there is no difference between EXISTS vs. IN, the query plan looks the same.
As there is no benefit should I change it to EXISTS?

BTW take a look 10 lines above ;)
https://github.com/xbmc/xbmc/pull/6941/files#diff-dab45208f50daf6a73c9472dd2ad7e3dL5749