@Montellese 2015-04-13
A while ago I started changing all IN conditions to EXISTS conditions in the videodb whenever I have to touch an existing query and I use EXISTS for all new queries. Are you sure that the query plan is the same? Is this for MySQL or SQLite?