@mkortstiege 2015-04-13
Nice! As for the view update you need to bump the db schema so it get's re-created.