@xhaggi 2015-04-14
@Montellese I've compared the query plan in sqlite and it differs for IN and EXISTS. As this has no effect to MySQL, but to SQLite i'll change it to EXISTS to be consistent with your changes in videodb.

@mkortstiege thanks for the hint, but i leave the version as it is because we want to test against the same database version. Will bump it now ;)